Create Cisco_Switch_Port_Scan.py
Basic switch scan for open (up) and closed (down) ports on a Cisco switch. The script itself is not  PURELY meant for Cisco switches, that's just what I have in my current environment. I plan to add in logic to automatically close an open and unused port to enhance port security.